+++ Libraries/XulbuX/src/XulbuX/xx_cmd.py
@@ -19,8 +19,10 @@ For more detailed information about formatting codes, see the `log` class descri
 """
 
 
-try: from .xx_format_codes import *
-except: from xx_format_codes import *
+from ._consts_ import DEFAULT
+from .xx_format_codes import *
+from .xx_string import *
+from .xx_color import *
 
 import keyboard as _keyboard
 import getpass as _getpass
@@ -29,6 +31,7 @@ import shutil as _shutil
 import msvcrt as _msvcrt
 import sys as _sys
 import os as _os
+
 
 
 
@@ -61,7 +64,7 @@ class Cmd:
             return _ctypes.windll.shell32.IsUserAnAdmin() in [1, True]
         except AttributeError:
             return False
-    
+
     @staticmethod
     def pause_exit(pause:bool = False, exit:bool = False, last_msg:str = '', exit_code:int = 0, reset_ansi:bool = False) -> None:
         print(last_msg, end='', flush=True)
@@ -78,50 +81,74 @@ class Cmd:
 
     @staticmethod
     def log(title:str, msg:str, start:str = '', end:str = '\n', title_bg_color:hexa|rgba = None, default_color:hexa|rgba = None) -> None:
+        """Will print a formatted log message:<br>
+        `title` -⠀the title of the log message (e.g. `DEBUG`, `WARN`, `FAIL`, etc.)<br>
+        `msg` -⠀the log message<br>
+        `start` -⠀something to print before the log is printed<br>
+        `end` -⠀something to print after the log is printed (e.g. `\\n\\n`)<br>
+        `title_bg_color` -⠀the background color of the `title`<br>
+        `default_color` -⠀the default text color of the `msg`<br>
+        The log message supports special formatting codes. For more detailed information about formatting codes, see `xx_format_codes` class description."""
         title_color = '_color' if not title_bg_color else Color.text_color_for_on_bg(title_bg_color)
         if title: FormatCodes.print(f'{start}  [bold][{title_color}]{f"[BG:{title_bg_color}]" if title_bg_color else ""} {title.upper()}: [_]\t{f"[{default_color}]" if default_color else ""}{str(msg)}[_]', default_color, end=end)
         else: FormatCodes.print(f'{start}  {f"[{default_color}]" if default_color else ""}{str(msg)}[_]', default_color, end=end)
 
     @staticmethod
     def debug(msg:str = 'Point in program reached.', active:bool = True, start:str = '\n', end:str = '\n\n', title_bg_color:hexa|rgba = DEFAULT.color['yellow'], default_color:hexa|rgba = DEFAULT.text_color, pause:bool = False, exit:bool = False) -> None:
+        """A preset for `log()`: `DEBUG` log message with the options to pause at the message and exit the program after the message was printed."""
         if active:
             Cmd.log('DEBUG', msg, start, end, title_bg_color, default_color)
             Cmd.pause_exit(pause, exit)
 
     @staticmethod
     def info(msg:str = 'Program running.', start:str = '\n', end:str = '\n\n', title_bg_color:hexa|rgba = DEFAULT.color['blue'], default_color:hexa|rgba = DEFAULT.text_color, pause:bool = False, exit:bool = False) -> None:
+        """A preset for `log()`: `INFO` log message with the options to pause at the message and exit the program after the message was printed."""
         Cmd.log('INFO', msg, start, end, title_bg_color, default_color)
         Cmd.pause_exit(pause, exit)
 
     @staticmethod
     def done(msg:str = 'Program finished.', start:str = '\n', end:str = '\n\n', title_bg_color:hexa|rgba = DEFAULT.color['teal'], default_color:hexa|rgba = DEFAULT.text_color, pause:bool = False, exit:bool = False) -> None:
+        """A preset for `log()`: `DONE` log message with the options to pause at the message and exit the program after the message was printed."""
         Cmd.log('DONE', msg, start, end, title_bg_color, default_color)
         Cmd.pause_exit(pause, exit)
 
     @staticmethod
     def warn(msg:str = 'Important message.', start:str = '\n', end:str = '\n\n', title_bg_color:hexa|rgba = DEFAULT.color['orange'], default_color:hexa|rgba = DEFAULT.text_color, pause:bool = False, exit:bool = False) -> None:
+        """A preset for `log()`: `WARN` log message with the options to pause at the message and exit the program after the message was printed."""
         Cmd.log('WARN', msg, start, end, title_bg_color, default_color)
         Cmd.pause_exit(pause, exit)
 
     @staticmethod
     def fail(msg:str = 'Program error.', start:str = '\n', end:str = '\n\n', title_bg_color:hexa|rgba = DEFAULT.color['red'], default_color:hexa|rgba = DEFAULT.text_color, pause:bool = False, exit:bool = True, reset_ansi=True) -> None:
+        """A preset for `log()`: `FAIL` log message with the options to pause at the message and exit the program after the message was printed."""
         Cmd.log('FAIL', msg, start, end, title_bg_color, default_color)
         Cmd.pause_exit(pause, exit, reset_ansi=reset_ansi)
 
     @staticmethod
     def exit(msg:str = 'Program ended.', start:str = '\n', end:str = '\n\n', title_bg_color:hexa|rgba = DEFAULT.color['magenta'], default_color:hexa|rgba = DEFAULT.text_color, pause:bool = False, exit:bool = True, reset_ansi=True) -> None:
+        """A preset for `log()`: `EXIT` log message with the options to pause at the message and exit the program after the message was printed."""
         Cmd.log('EXIT', msg, start, end, title_bg_color, default_color)
         Cmd.pause_exit(pause, exit, reset_ansi=reset_ansi)
 
     @staticmethod
-    def confirm(msg:str = 'Are you sure? [_|dim]((Y/n):  )', start = '\n', end = '\n', default_color:hexa|rgba = DEFAULT.color['cyan'], default_is_yes:bool = True) -> None:
-        confirmed = input(FormatCodes.to_ansi(f'{start}  {str(msg)}', default_color)).strip().lower() in (('', 'y', 'yes') if default_is_yes else ('y', 'yes'))
+    def confirm(msg:str = 'Do you want to continue?', start = '\n', end = '\n', default_color:hexa|rgba = DEFAULT.color['cyan'], default_is_yes:bool = True) -> None:
+        """Ask a yes/no question.\n
+        -----------------------------------------------------------------------------------
+        The question can be formatted with special formatting codes. For more detailed<br>
+        information about formatting codes, see `xx_format_codes` class description."""
+        confirmed = input(FormatCodes.to_ansi(f'{start}  {str(msg)} [_|dim](({"Y" if default_is_yes else "y"}/{"n" if default_is_yes else "N"}):  )', default_color)).strip().lower() in (('', 'y', 'yes') if default_is_yes else ('y', 'yes'))
         if end: Cmd.log('', '') if end == '\n' else Cmd.log('', end[1:]) if end.startswith('\n') else Cmd.log('', end)
         return confirmed
 
     @staticmethod
-    def restricted_input(prompt:str = '', allowed_chars:str = '0123456789', min_length:int = None, max_length:int = None):
-        print(prompt, end='', flush=True)
+    def input(prompt:object = '', allowed_chars:str = '0123456789', min_length:int = None, max_length:int = None) -> str:
+        """Acts like a standard Python `input()` with the advantage, that you can specify:
+        - what text characters the user is allowed to type and
+        - the minimum and/or maximum length of the users input.\n
+        -----------------------------------------------------------------------------------
+        The input can be formatted with special formatting codes. For more detailed<br>
+        information about formatting codes, see `xx_format_codes` class description."""
+        print(FormatCodes.to_ansi(prompt), end='', flush=True)
         result = ''
         while True:
             char = _msvcrt.getch().decode('utf-8', errors='ignore')