+++ Libraries/XulbuX/src/XulbuX/xx_data.py
@@ -3,6 +3,7 @@ import re as _re
 
 
 
+
 class Data:
 
     @staticmethod