+++ Libraries/XulbuX/src/XulbuX/xx_string.py
@@ -2,6 +2,7 @@ import re as _re
 
 
 
+
 class String:
 
     @staticmethod